[AC-6074] clean-up permissions on impact, move every permission into a file in a single 'permissions' directory
