+++ web/impact/impact/permissions/allocate_applications_permissions.py
