+++ web/impact/impact/permissions/directory_access_permissions.py
@@ -1,0 +1,17 @@
+from impact.permissions import BasePermission
+from accelerator_abstract.models import ACTIVE_PROGRAM_STATUS
+from accelerator.models import UserRole
+from accelerator_abstract.models.base_user_utils import is_employee
+
+
+class DirectoryAccessPermissions(BasePermission):
+    authenticated_users_only = True
+
+    def has_permission(self, request, view):
+        allowed_roles = UserRole.FINALIST_USER_ROLES + [
+            UserRole.MENTOR, UserRole.ALUM
+        ]
+        return request.user.programrolegrant_set.filter(
+            program_role__program__program_status=ACTIVE_PROGRAM_STATUS,
+            program_role__user_role__name__in=allowed_roles,
+        ).exists() or is_employee(request.user)
