+++ web/impact/impact/permissions/dynamic_model_permissions.py
@@ -1,33 +1,12 @@
-# MIT License
-# Copyright (c) 2017 MassChallenge, Inc.
-
-from ast import literal_eval
-import logging
-
-from django.conf import settings
-from django.contrib.auth import get_user
-from django.contrib.auth import get_user_model
+from impact.permissions import BasePermission
 from django.contrib.auth.models import Permission
-from django.contrib.auth.models import PermissionDenied
 from django.contrib.contenttypes.models import ContentType
-from rest_framework.permissions import BasePermission
-
 from accelerator.apps import AcceleratorConfig
-from accelerator.models import Clearance
-from accelerator_abstract.models.base_clearance import (
-    CLEARANCE_LEVEL_GLOBAL_MANAGER,
-    CLEARANCE_LOGGER_FAILED_INSUFFICIENT_CLEARANCE_MSG,
-    CLEARANCE_LOGGER_SUCCESS_MSG,
-)
 from impact.utils import model_name_case
+from django.contrib.auth import get_user
+from django.contrib.auth.models import PermissionDenied
+from ast import literal_eval
 
-from accelerator_abstract.models.base_user_utils import is_entrepreneur
-from accelerator_abstract.models import ACTIVE_PROGRAM_STATUS
-from accelerator.models import UserRole
-from accelerator_abstract.models.base_user_utils import is_employee
-
-logger = logging.getLogger(__file__)
-User = get_user_model()
 
 METHOD_TO_ACTION = {
     "POST": "add",
@@ -38,72 +17,6 @@ METHOD_TO_ACTION = {
     "PATCH": "change",
     "DELETE": "delete",
 }
-
-
-# FIXME: function definitions duplicated from mc.permission_checks
-
-def _log_access_attempt(cleared, user, level, program_family):
-    if cleared:
-        _log_sufficient_clearance(level, program_family, user)
-    else:
-        _log_insufficient_clearance(level, program_family, user)
-
-
-def _log_sufficient_clearance(level, program_family, user):
-    logger.info(CLEARANCE_LOGGER_SUCCESS_MSG.format(
-        user=user, program_family=program_family, level=level))
-
-
-def _log_insufficient_clearance(level, program_family, user):
-    logger.info(CLEARANCE_LOGGER_FAILED_INSUFFICIENT_CLEARANCE_MSG.format(
-        user=user, program_family=program_family, level=level))
-
-# end duplicated definitions
-
-
-def global_operations_manager_check(user, program_family=None):
-    cleared = Clearance.objects.check_clearance(
-        user, CLEARANCE_LEVEL_GLOBAL_MANAGER, program_family)
-    _log_access_attempt(cleared, user,
-                        CLEARANCE_LEVEL_GLOBAL_MANAGER, program_family)
-    return cleared
-
-
-class V0APIPermissions(BasePermission):
-    authenticated_users_only = True
-
-    def has_permission(self, request, view):
-        return request.user.groups.filter(
-            name=settings.V0_API_GROUP).exists()
-
-
-class DirectoryAccessPermissions(BasePermission):
-    authenticated_users_only = True
-
-    def has_permission(self, request, view):
-        allowed_roles = UserRole.FINALIST_USER_ROLES + [
-            UserRole.MENTOR, UserRole.ALUM
-        ]
-        return request.user.programrolegrant_set.filter(
-            program_role__program__program_status=ACTIVE_PROGRAM_STATUS,
-            program_role__user_role__name__in=allowed_roles,
-        ).exists() or is_employee(request.user)
-
-
-class V1APIPermissions(BasePermission):
-    authenticated_users_only = True
-
-    def has_permission(self, request, view):
-        return request.user.groups.filter(
-            name=settings.V1_API_GROUP).exists()
-
-
-class V1ConfidentialAPIPermissions(BasePermission):
-    authenticated_users_only = True
-
-    def has_permission(self, request, view):
-        return request.user.groups.filter(
-            name=settings.V1_CONFIDENTIAL_API_GROUP).exists()
 
 
 def method_to_perm(method):
